+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - (API) Added `ChunkService#removeUnusedChunksImmediately` method that can remove unused chunks immediately. Also, the `/gc` command
   will call this method in all dimensions now.
 - (API) Added `ItemBaseComponent#getLockMode` and `ItemBaseComponent#setLockMode` methods to get and set the lock mode of an item.
+- Implemented reeds (also called sugar cane).
 
 ### Changed